+++ cpp/include/tudat-learn/estimators/regressor.hpp
@@ -63,7 +63,7 @@ class Regressor : public Estimator<Datum_t, Label_t> {
      * 
      * @param fit_indices Vector with the indices of the feature vectors to which the Regressor is going to be fitted.
      */
-    virtual void fit(const std::vector<int> &fit_indices) = 0;
+    virtual void fit(const std::vector<size_t> &fit_indices) = 0;
 
 };
